requirements.txt
Django==2.2.3
django-excel==0.0.10
django-pure-pagination==0.3.0
lml==0.0.9
mysqlclient==1.4.2.post1  # 如果是部署在linux上面, 则将其改为mysqlclient==1.4.2
pyexcel==0.5.15
pyexcel-io==0.5.20
pyexcel-webio==0.1.4
pyexcel-xls==0.5.8
pytz==2019.1
sqlparse==0.3.0
texttable==1.6.2
xlrd==1.2.0
xlwt==1.3.0
